Separate voteDialog and voteDialogHOC
