+++ src/components/voteDialog/voteDialog.css
@@ -1,0 +1,5 @@
+.info {
+    border-top: 1px solid rgba(0,0,0,0.12);
+    border-bottom: 1px solid rgba(0,0,0,0.12);
+    margin: 10px -24px 20px;
+}
